requirements.txt
ccxt==4.2.36
fastapi==0.85.2
numpy
pandas==2.1.3
pymongo==4.3.2
python-dotenv==0.21.0
uvicorn==0.19.0
pyyaml==5.3.1
dask[complete]
pytest==7.4.2
nose==1.3.7
autopylogger==1.0
boto3